Renamed MainMenu and SettingsMenu
