+++ Assets/Prefabs/PickupPup/Core/UI/MainMenuUI.prefab
@@ -1273,8 +1273,7 @@ MonoBehaviour:
   preserveOnSceneChange: 0
   alternateSprites: []
   shouldCollectInSpawnPool: 1
-  overrideAutoSort: 0
-  isTopLayerCanvas: 0
+  sortRule: 0
 --- !u!114 &114589268010027904
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -1822,7 +1821,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1302222667971700}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f3036b15e30914e8bade48d4902569f7, type: 3}
+  m_Script: {fileID: 11500000, guid: 71f8b7ec12bb1e045adbba136d45be01, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   preserveOnSceneChange: 0
