+++ Assets/Prefabs/PickupPup/Core/UI/SettingsMenu.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 36addd15c2c3143fca398d5e62972863
-timeCreated: 1484826905
+guid: bb45bc5c245b0744eb14f3243080f954
+timeCreated: 1486656200
 licenseType: Free
 NativeFormatImporter:
   userData:
