+++ Assets/Scripts/UI/PickupPup/MainMenu/MainMenu.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 0a5fe1dc07dfa4298986bf1b9f9af6b2
-timeCreated: 1484826023
+guid: 71f8b7ec12bb1e045adbba136d45be01
+timeCreated: 1486655921
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
