+++ Assets/Scripts/UI/PickupPup/SettingsMenu.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 029b56c77ee044119826b513a3cfcda6
+guid: 4bfd0210bba4d4642a32cd755443e2a7
 folderAsset: yes
-timeCreated: 1484825993
+timeCreated: 1486656062
 licenseType: Free
 DefaultImporter:
   userData:
