+++ Assets/Scripts/UI/PickupPup/SettingsMenu/SettingsMenu.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: f3036b15e30914e8bade48d4902569f7
-timeCreated: 1484457786
+guid: 53bbaa2356d8c5c468cc446aa50f7c16
+timeCreated: 1486656100
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
